leadgen-ai/requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml

# Browser Automation (optional for MVP)
# playwright>=1.42.0

# Database
sqlalchemy>=2.0.0

# Data Processing
# pandas>=2.2.0

# AI Integration (new unified Google GenAI SDK)
google-genai>=1.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Utilities
validators>=0.22.0
